+++ enterprise/server/include/sourcemeta/one/enterprise_server_actions.h
@@ -6,49 +6,35 @@
 
 #include <sourcemeta/blaze/evaluator.h>
 
+#include <sourcemeta/one/dispatcher.h>
 #include <sourcemeta/one/http.h>
 
 #include <filesystem>  // std::filesystem::path
-#include <functional>  // std::function
-#include <string>      // std::string
+#include <span>        // std::span
 #include <string_view> // std::string_view
 
-class EnterpriseMCP {
+namespace sourcemeta::one::enterprise {
+
+class ActionMCP_v1 : public sourcemeta::one::Action {
 public:
-  // Handler invoked on `tools/call`. Takes the raw JSON-RPC envelope
-  // and the loaded `mcp.metapack`, returns the JSON-RPC response
-  // envelope. Set by the owning `ActionMCP_v1` (which has access to
-  // the `ActionDispatcher` and can route to sibling actions' `mcp()`
-  // methods). The hook is the dependency-inversion seam that lets
-  // `enterprise_server` avoid linking against `actions`
-  using ToolCallHandler = std::function<sourcemeta::core::JSON(
-      const sourcemeta::core::JSON &envelope,
-      const sourcemeta::core::JSON &mcp_metadata)>;
+  static constexpr std::string_view DESCRIPTION{
+      "Handle Model Context Protocol JSON-RPC requests"};
 
-  EnterpriseMCP(const std::filesystem::path &base,
-                const sourcemeta::core::URITemplateRouterView &router,
-                sourcemeta::core::URITemplateRouter::Identifier identifier);
+  ActionMCP_v1(const std::filesystem::path &base,
+               const sourcemeta::core::URITemplateRouterView &router,
+               sourcemeta::core::URITemplateRouter::Identifier identifier);
 
-  // To avoid mistakes
-  EnterpriseMCP(const EnterpriseMCP &) = delete;
-  EnterpriseMCP(EnterpriseMCP &&) = delete;
-  auto operator=(const EnterpriseMCP &) -> EnterpriseMCP & = delete;
-  auto operator=(EnterpriseMCP &&) -> EnterpriseMCP & = delete;
-  ~EnterpriseMCP() = default;
-
-  auto set_tool_call_handler(ToolCallHandler handler) -> void;
-
-  auto rest(sourcemeta::one::HTTPRequest &request,
-            sourcemeta::one::HTTPResponse &response) -> void;
+  auto rest(const std::span<std::string_view> matches,
+            sourcemeta::one::HTTPRequest &request,
+            sourcemeta::one::HTTPResponse &response) -> void override;
 
 private:
-  std::filesystem::path base_;
   std::string_view allowed_origin_;
-  std::string_view registry_url_;
   std::string_view response_schema_;
   sourcemeta::blaze::Template request_schema_template_;
   sourcemeta::core::JSON mcp_metadata_{nullptr};
-  ToolCallHandler tool_call_handler_;
 };
+
+} // namespace sourcemeta::one::enterprise
 
 #endif
