+++ src/actions/action_jsonschema_serve_v1.h
@@ -3,7 +3,7 @@
 
 #include <sourcemeta/core/uritemplate.h>
 
-#include <sourcemeta/one/actions.h>
+#include <sourcemeta/one/dispatcher.h>
 #include <sourcemeta/one/http.h>
 #include <sourcemeta/one/shared.h>
