+++ src/actions/action_mcp_v1.h
@@ -2,61 +2,10 @@
 #define SOURCEMETA_ONE_ACTIONS_MCP_V1_H
 
 #if defined(SOURCEMETA_ONE_ENTERPRISE)
-#include <sourcemeta/core/uritemplate.h>
 
-#include <sourcemeta/one/actions.h>
 #include <sourcemeta/one/enterprise_server.h>
-#include <sourcemeta/one/http.h>
-#include <sourcemeta/one/mcp.h>
 
-#include <exception>   // std::exception
-#include <filesystem>  // std::filesystem
-#include <span>        // std::span
-#include <string_view> // std::string_view
-
-class ActionMCP_v1 : public sourcemeta::one::Action, public EnterpriseMCP {
-public:
-  static constexpr std::string_view DESCRIPTION{
-      "Handle Model Context Protocol JSON-RPC requests"};
-
-  ActionMCP_v1(const std::filesystem::path &base,
-               const sourcemeta::core::URITemplateRouterView &router,
-               const sourcemeta::core::URITemplateRouter::Identifier identifier)
-      : sourcemeta::one::Action{base, router.base_path(), router.base_url()},
-        EnterpriseMCP{base, router, identifier} {}
-
-  auto rest(const std::span<std::string_view>,
-            sourcemeta::one::HTTPRequest &request,
-            sourcemeta::one::HTTPResponse &response) -> void override {
-    EnterpriseMCP::rest(request, response);
-  }
-
-  auto set_dispatcher(sourcemeta::one::ActionDispatcher &dispatcher)
-      -> void override {
-    EnterpriseMCP::set_tool_call_handler(
-        [&dispatcher](const sourcemeta::core::JSON &envelope,
-                      const sourcemeta::core::JSON &mcp_metadata)
-            -> sourcemeta::core::JSON {
-          const auto &id{envelope.at("id")};
-          const auto &name{envelope.at("params").at("name").to_string()};
-          const auto &tool_routes{mcp_metadata.at("toolRoutes")};
-          if (!tool_routes.defines(name)) {
-            return sourcemeta::core::jsonrpc_make_error_invalid_params(
-                id, sourcemeta::core::JSON{name});
-          }
-          const auto identifier{
-              static_cast<sourcemeta::core::URITemplateRouter::Identifier>(
-                  tool_routes.at(name).to_integer())};
-          auto *instance{dispatcher.action(identifier)};
-          assert(instance != nullptr);
-          try {
-            return instance->mcp(envelope);
-          } catch (const std::exception &error) {
-            return sourcemeta::one::mcp_make_tool_error(id, error.what());
-          }
-        });
-  }
-};
+using ActionMCP_v1 = sourcemeta::one::enterprise::ActionMCP_v1;
 
 #else
 
@@ -65,7 +14,7 @@ public:
 #include <sourcemeta/core/uri.h>
 #include <sourcemeta/core/uritemplate.h>
 
-#include <sourcemeta/one/actions.h>
+#include <sourcemeta/one/dispatcher.h>
 #include <sourcemeta/one/http.h>
 #include <sourcemeta/one/metapack.h>
 #include <sourcemeta/one/shared.h>
