+++ src/actions/action_not_found_v1.h
@@ -3,7 +3,7 @@
 
 #include <sourcemeta/core/uritemplate.h>
 
-#include <sourcemeta/one/actions.h>
+#include <sourcemeta/one/dispatcher.h>
 #include <sourcemeta/one/http.h>
 
 #include <filesystem>  // std::filesystem
