--- src/actions/dispatch.cc
@@ -1,132 +1,0 @@
-#include <sourcemeta/one/actions.h>
-
-#include "action_default_v1.h"
-#include "action_health_check_v1.h"
-#include "action_jsonschema_evaluate_v1.h"
-#include "action_jsonschema_trace_v1.h"
-#include "action_mcp_v1.h"
-#include "action_not_found_v1.h"
-#include "action_schema_search_v1.h"
-#include "action_serve_explorer_artifact_v1.h"
-#include "action_serve_schema_artifact_v1.h"
-#include "action_serve_static_v1.h"
-
-#include <array>  // std::array
-#include <memory> // std::unique_ptr, std::make_unique
-#include <mutex>  // std::once_flag, std::call_once
-#include <string> // std::string
-
-template <typename T>
-static auto
-make_action(const std::filesystem::path &base,
-            const sourcemeta::core::URITemplateRouterView &router,
-            const sourcemeta::core::URITemplateRouter::Identifier identifier)
-    -> std::unique_ptr<sourcemeta::one::Action> {
-  return std::make_unique<T>(base, router, identifier);
-}
-
-using ActionConstructFunction =
-    auto (*)(const std::filesystem::path &,
-             const sourcemeta::core::URITemplateRouterView &,
-             sourcemeta::core::URITemplateRouter::Identifier)
-        -> std::unique_ptr<sourcemeta::one::Action>;
-
-#define SOURCEMETA_ONE_MAKE_CONSTRUCTOR_ENTRY(Name, Class)                     \
-  table[sourcemeta::one::ACTION_TYPE_##Name] = &make_action<Class>;
-
-static constexpr std::array<ActionConstructFunction,
-                            sourcemeta::one::ACTION_TYPE_COUNT>
-    CONSTRUCTORS{[] {
-      std::array<ActionConstructFunction, sourcemeta::one::ACTION_TYPE_COUNT>
-          table{};
-      SOURCEMETA_ONE_FOR_EACH_ACTION(SOURCEMETA_ONE_MAKE_CONSTRUCTOR_ENTRY)
-      return table;
-    }()};
-
-#undef SOURCEMETA_ONE_MAKE_CONSTRUCTOR_ENTRY
-
-#define SOURCEMETA_ONE_DEFINE_DESCRIPTION(Name, Class) Class::DESCRIPTION,
-
-static constexpr std::array<std::string_view,
-                            sourcemeta::one::ACTION_TYPE_COUNT>
-    DESCRIPTIONS{
-        {SOURCEMETA_ONE_FOR_EACH_ACTION(SOURCEMETA_ONE_DEFINE_DESCRIPTION)}};
-
-#undef SOURCEMETA_ONE_DEFINE_DESCRIPTION
-
-auto sourcemeta::one::ActionDispatcher::description(
-    const sourcemeta::core::URITemplateRouter::Identifier context) noexcept
-    -> std::string_view {
-  if (context >= DESCRIPTIONS.size()) {
-    return {};
-  }
-  return DESCRIPTIONS[context];
-}
-
-sourcemeta::one::ActionDispatcher::ActionDispatcher(
-    const std::filesystem::path &base,
-    const sourcemeta::core::URITemplateRouterView &router)
-    : base_{base}, router_{router},
-      // NOLINTNEXTLINE(modernize-avoid-c-arrays)
-      slots_{std::make_unique<Slot[]>(router.size() + 1)},
-      slots_size_{router.size() + 1} {
-  router.arguments(0, [this](const auto &key, const auto &value) {
-    if (key == "errorSchema") {
-      this->default_error_schema_ = std::get<std::string_view>(value);
-    }
-  });
-}
-
-auto sourcemeta::one::ActionDispatcher::error(
-    const sourcemeta::one::HTTPRequest &request,
-    sourcemeta::one::HTTPResponse &response, const char *const code,
-    std::string &&identifier, std::string &&message) const -> void {
-  sourcemeta::one::json_error(request, response, code, std::move(identifier),
-                              std::move(message), this->default_error_schema_);
-}
-
-auto sourcemeta::one::ActionDispatcher::action(
-    const sourcemeta::core::URITemplateRouter::Identifier identifier,
-    const sourcemeta::core::URITemplateRouter::Identifier context)
-    -> sourcemeta::one::Action * {
-  if (identifier >= this->slots_size_ || context >= CONSTRUCTORS.size())
-      [[unlikely]] {
-    return nullptr;
-  }
-
-  auto &slot{this->slots_[identifier]};
-  std::call_once(slot.flag, [this, &slot, context, identifier] {
-    slot.instance =
-        CONSTRUCTORS[context](this->base_, this->router_, identifier);
-    slot.instance->set_dispatcher(*this);
-  });
-
-  return slot.instance.get();
-}
-
-auto sourcemeta::one::ActionDispatcher::action(
-    const sourcemeta::core::URITemplateRouter::Identifier identifier)
-    -> sourcemeta::one::Action * {
-  if (identifier >= this->slots_size_) [[unlikely]] {
-    return nullptr;
-  }
-  return this->action(identifier, this->router_.context(identifier));
-}
-
-auto sourcemeta::one::ActionDispatcher::dispatch(
-    const sourcemeta::core::URITemplateRouter::Identifier identifier,
-    const sourcemeta::core::URITemplateRouter::Identifier context,
-    const std::span<std::string_view> matches,
-    sourcemeta::one::HTTPRequest &request,
-    sourcemeta::one::HTTPResponse &response) -> void {
-  auto *instance{this->action(identifier, context)};
-  if (instance == nullptr) [[unlikely]] {
-    this->error(request, response, sourcemeta::one::STATUS_NOT_IMPLEMENTED,
-                "unknown-handler-code",
-                "This server version does not implement the handler for "
-                "this URL");
-    return;
-  }
-
-  instance->rest(matches, request, response);
-}
